Changes in bootstrap to support parameters.
Paths to a test project can be passed by adding ?"path1"|"path2"|"path3"
after ,SetJavaBenchmark.
The | allows you to add more paths but this is optional.

Software units can be mapped by adding
?"softwareUnit1">"PACKAGE"|"softwareUnit"|"SoftwareUnit">"type2".
Here you can use the | as a separator as well.
The > is optional. It maps the type, however if left out it will use
"PACKAGE" by default.